Add CORS setup instructions to README
WorkOS Widgets requires that your web origin be allowed
in order to function. Failure to add your domain to web
origins that are allowed will result in ambiguous loading
errors.

This commit updates the README to include the instruction
and also adds the NEXT_PUBLIC_WORKOS_REDIRECT_URI to the
suggested .env.local file.